replaced dashboard with newdashboard
