--- client/components/Chart.tsx
@@ -1,5 +1,0 @@
-/**
- * Apex chart may go here
- */
-
-export {};
