+++ client/components/Dashboard.tsx
@@ -1,434 +1,53 @@
-export {};
-// import React, { useState, useEffect } from 'react';
-// import { useQuery, gql } from '@apollo/client';
-// import { useNavigate } from 'react-router';
-// import Chart from 'react-apexcharts';
-// import ChartBox from './ChartBox';
-// import { SelectedProject, ProjectQuery } from '../../@types/Interfaces';
-// import { SortOrder, ChartData } from '../../@types/dashboard';
-// import Queries from './Queries';
-// import { useAuth } from '../auth/AuthProvider';
+import React, { useState, useEffect } from 'react';
+// import { gql } from 'apollo-server-express';
+import { useQuery, gql } from '@apollo/client';
+// import { useQuery } from '@apollo/client';
+import ToolBar from './ToolBar';
+import ProjectView from './ProjectView';
+import { useAuth } from '../auth/AuthProvider';
+// import { SelectedProject, Projects } from '../../@types/Interfaces';
 
-// const GET_USER_DATA = gql`
-//     query getUserData($userId: String!) {
-//         user(id: $userId) {
-//             # note this curretnly gets all projects and queries. This could be a very expersive query.
-//             projects {
-//                 name
-//                 id
-//                 queries {
-//                     number
-//                     latency
-//                     complexity
-//                     depth
-//                     timestamp
-//                     tokens
-//                     success
-//                 }
-//             }
-//         }
-//     }
-// `;
+const GET_PROJECT_DATA = gql`
+    query getUserData($userId: String!) {
+        user(id: $userId) {
+            projects {
+                name
+                id
+                userID
+                apiKey
+            }
+        }
+    }
+`;
 
-// // const GET_PROJECT = gql`
-// //     query Query($projectId: String!) {
-// //         project(id: $projectId) {
-// //             id
-// //             queries {
-// //                 timestamp
-// //                 depth
-// //                 complexity
-// //             }
-// //         }
-// //     }
-// // `;
+export default function Dashboard() {
+    /** Bring the user context into this component */
+    const { user } = useAuth();
 
-// // we are defining state types - arrow (for sorting arrow change when clicked -ascending or descending) and time,depth and complexity props of a query
-// export interface ISState {
-//     style: {
-//         time: boolean; // TODO: Change to timestamp
-//         depth: boolean;
-//         complexity: boolean;
-//     };
-//     arrow: {
-//         timestamp: SortOrder;
-//         depth: SortOrder;
-//         complexity: SortOrder;
-//     };
-// }
+    /** State requirments for this component */
+    const [selectedProject, setSelectedProject] = useState<Project>();
+    const [allProjects, setAllProjects] = useState<Project[]>();
 
-// function Dashboard() {
-//     /** Bring the user context into this component */
-//     const { user } = useAuth();
+    /** Send query to get project information for this user */
+    const { data, loading } = useQuery(GET_PROJECT_DATA, {
+        variables: {
+            userId: user!.id,
+        },
+    });
+    useEffect(() => {
+        if (!loading && data) {
+            setAllProjects(data.user.projects);
+        }
+    }, [loading, data]);
 
-//     /** Layout the state requirments for this component */
-//     const [style, setStyle] = useState<ISState['style']>({
-//         time: false,
-//         depth: false,
-//         complexity: false,
-//     });
-//     const [arrow, setArrow] = useState<ISState['arrow']>({
-//         // adjusting arrow based on ascending vs descending
-//         timestamp: '',
-//         depth: '',
-//         complexity: '',
-//     });
-//     const [setProject, selectProject] = useState<SelectedProject['project']>();
-//     const [queries, setQueries] = useState<ProjectQuery[]>();
-//     // const [projects, setProjects] = useState<Projects['projects']>();   // commented out due to still not knowing how you guys want our project selection to look like. Waiting for instructions
-//     // const [order, setOrder] = useState<boolean>(); // ascending vs descending on click
-
-//     /** Layout the options and state requirements for the chart */
-//     const [chartData, setChartData] = useState<ChartData | null>(null);
-
-//     /** Query the databes for project data */
-//     const { data, loading } = useQuery(GET_USER_DATA, {
-//         variables: {
-//             userId: user!.id,
-//         },
-//     });
-//     // const { data, loading } = useQuery(GET_PROJECT, {
-//     //     variables: {
-//     //         projectId: dataR.projects.id,
-//     //     },
-//     // });
-
-//     // toggle classes on click in the Queries component
-
-//     const setToggle = (arg: string) => {
-//         if (arg === 'time') {
-//             setStyle({
-//                 ...style,
-//                 time: true,
-//                 depth: false,
-//                 complexity: false,
-//             });
-//         } else if (arg === 'depth') {
-//             setStyle({
-//                 ...style,
-//                 time: false,
-//                 depth: true,
-//                 complexity: false,
-//             });
-//         } else if (arg === 'complexity') {
-//             setStyle({
-//                 ...style,
-//                 time: false,
-//                 depth: false,
-//                 complexity: true,
-//             });
-//         }
-//     };
-
-//     // const test = (pr: any): void => { // this is poorly named fuction that selects the project
-//     //     selectProject(pr);
-//     // };
-
-//     // sorting functions
-
-//     const combinedSort = (field: keyof ISState['arrow'], sortOrder: SortOrder): void => {
-//         if (setProject) {
-//             const newArr = [];
-//             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//             for (const key in setProject?.queries) {
-//                 newArr.push(setProject?.queries[key]);
-//             }
-//             newArr.sort((a, b) => {
-//                 if (sortOrder === '↑') {
-//                     return a[field] - b[field];
-//                 }
-//                 if (sortOrder === '↓') {
-//                     return b[field] - a[field];
-//                 }
-//                 return 0;
-//             });
-//             setArrow({
-//                 timestamp: '',
-//                 depth: '',
-//                 complexity: '',
-//                 [field]: sortOrder,
-//             });
-//             setQueries(newArr);
-//         }
-//     };
-
-//     // const sortByTime = (): void => {
-//     //     if (order === true) {
-//     //         setOrder(false);
-//     //         // setArrow();
-//     //         setArrow({
-//     //             ...arrow,
-//     //             time: '↑',
-//     //             depth: '',
-//     //             complexity: '',
-//     //         });
-//     //     } else {
-//     //         setOrder(true);
-//     //         setArrow({
-//     //             ...arrow,
-//     //             time: '↓',
-//     //             depth: '',
-//     //             complexity: '',
-//     //         });
-//     //     }
-//     //     if (order) {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => a.timestamp - b.timestamp);
-//     //         setQueries(dataToSort);
-//     //     } else {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => a.timestamp + b.timestamp);
-//     //         setQueries(dataToSort);
-//     //     }
-//     // };
-//     // const sortByDepth = (): void => {
-//     //     if (order === true) {
-//     //         setOrder(false);
-//     //         setArrow({
-//     //             ...arrow,
-//     //             depth: '↑',
-//     //             time: '',
-//     //             complexity: '',
-//     //         });
-//     //     } else {
-//     //         setOrder(true);
-//     //         setArrow({
-//     //             ...arrow,
-//     //             depth: '↓',
-//     //             time: '',
-//     //             complexity: '',
-//     //         });
-//     //     }
-//     //     if (order) {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => a.depth - b.depth);
-//     //         setQueries(dataToSort);
-//     //     } else {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => b.depth - a.depth);
-//     //         setQueries(dataToSort);
-//     //     }
-//     // };
-//     // const sortByComplexity = (): void => {
-//     //     if (order === true) {
-//     //         setOrder(false);
-//     //         setArrow({
-//     //             ...arrow,
-//     //             complexity: '↑',
-//     //             time: '',
-//     //             depth: '',
-//     //         });
-//     //     } else {
-//     //         setOrder(true);
-//     //         setArrow({
-//     //             ...arrow,
-//     //             complexity: '↓',
-//     //             depth: '',
-//     //             time: '',
-//     //         });
-//     //     }
-//     //     if (order) {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => a.complexity - b.complexity);
-//     //         setQueries(dataToSort);
-//     //     } else {
-//     //         const newArr = [];
-//     //         if (selectedProject) {
-//     //             // eslint-disable-next-line no-restricted-syntax, guard-for-in
-//     //             for (const key in selectedProject?.queries) {
-//     //                 newArr.push(selectedProject?.queries[key]);
-//     //             }
-//     //         }
-//     //         const dataToSort = [...newArr];
-//     //         dataToSort.sort((a, b) => b.complexity - a.complexity);
-//     //         setQueries(dataToSort);
-//     //     }
-//     // };
-
-//     useEffect(() => {
-//         // if (!loadingR && dataR) {
-//         //     setProjects(dataR.user.projects);
-//         // }
-//         if (!loading && data) {
-//             selectProject(data.user.projects[0]);
-//             setQueries(setProject?.queries);
-//         }
-//     }, [/** loadingR, dataR, */ loading, data, setProject?.queries]);
-
-//     return (
-//         <div id="dashWrapper">
-//             <div className="loggerBox">
-//                 <div className="loggerSortButtonsWrapper">
-//                     <div className="loggerSortButtons">
-//                         {
-//                             // const Querries: React.FC<IProps> = ({ projects, test, sortByNameAsc, queries }) => (}
-//                         }
-//                         <div
-//                             aria-hidden="true"
-//                             className={`loggerBtn${style.time ? ' active' : ''}`}
-//                             onClick={() => {
-//                                 // sortByTime();
-//                                 let sortOrder: SortOrder;
-//                                 if (arrow.timestamp === '' || arrow.timestamp === '↓') {
-//                                     sortOrder = '↑';
-//                                 } else {
-//                                     sortOrder = '↓';
-//                                 }
-
-//                                 combinedSort('timestamp', sortOrder);
-//                                 setToggle('time');
-//                             }}
-//                         >
-//                             Time {arrow.timestamp}
-//                         </div>
-//                         <div
-//                             aria-hidden="true"
-//                             className={`loggerBtn${style.depth ? ' active' : ''}`}
-//                             onClick={() => {
-//                                 // sortByDepth();
-//                                 let sortOrder: SortOrder;
-//                                 if (arrow.depth === '' || arrow.depth === '↓') {
-//                                     sortOrder = '↑';
-//                                 } else {
-//                                     sortOrder = '↓';
-//                                 }
-//                                 combinedSort('depth', sortOrder);
-//                                 setToggle('depth');
-//                             }}
-//                         >
-//                             Depth {arrow.depth}
-//                         </div>
-//                         <div
-//                             aria-hidden="true"
-//                             className={`loggerBtn${arrow.complexity ? ' active' : ''}`}
-//                             onClick={() => {
-//                                 // sortByComplexity();
-//                                 let sortOrder: SortOrder;
-//                                 if (arrow.complexity === '' || arrow.complexity === '↓') {
-//                                     sortOrder = '↑';
-//                                 } else {
-//                                     sortOrder = '↓';
-//                                 }
-//                                 combinedSort('complexity', sortOrder);
-//                                 setToggle('complexity');
-//                             }}
-//                         >
-//                             Complexity {arrow.complexity}
-//                         </div>
-//                     </div>
-//                 </div>
-//                 {/* <div
-//                     aria-hidden="true"
-//                     className={`loggerBtn${style.time ? ' active' : ''}`}
-//                     onClick={() => {
-//                         // sortByTime();
-//                         let sortOrder: SortOrder;
-//                         if (arrow.timestamp === '' || arrow.timestamp === '↓') {
-//                             sortOrder = '↑';
-//                         } else {
-//                             sortOrder = '↓';
-//                         }
-
-//                         combinedSort('timestamp', sortOrder);
-//                         setToggle('time');
-//                     }}
-//                 >
-//                     Time {arrow.timestamp}
-//                 </div>
-//                 <div
-//                     aria-hidden="true"
-//                     className={`loggerBtn${style.depth ? ' active' : ''}`}
-//                     onClick={() => {
-//                         // sortByDepth();
-//                         let sortOrder: SortOrder;
-//                         if (arrow.depth === '' || arrow.depth === '↓') {
-//                             sortOrder = '↑';
-//                         } else {
-//                             sortOrder = '↓';
-//                         }
-//                         combinedSort('depth', sortOrder);
-//                         setToggle('depth');
-//                     }}
-//                 >
-//                     Depth {arrow.depth}
-//                 </div>
-//                 <div
-//                     aria-hidden="true"
-//                     className={`loggerBtn${style.complexity ? ' active' : ''}`}
-//                     onClick={() => {
-//                         // sortByComplexity();
-//                         let sortOrder: SortOrder;
-//                         if (arrow.complexity === '' || arrow.complexity === '↓') {
-//                             sortOrder = '↑';
-//                         } else {
-//                             sortOrder = '↓';
-//                         }
-//                         combinedSort('complexity', sortOrder);
-//                         setToggle('complexity');
-//                     }}
-//                 >
-//                     Complexity {arrow.complexity}
-//                 </div> */}
-//                 <div className="loggerGUI">
-//                     {/* <Logger test={test} projects={projects} /> */}
-//                     {/* <Querries queries={queries} sortByNameAsc={sortByNameAsc} test={test} projects={projects} /> */}
-//                     <Queries
-//                         arrowTime={arrow.timestamp}
-//                         arrowDepth={arrow.depth}
-//                         arrowComplexity={arrow.complexity}
-//                         time={style.time}
-//                         depth={style.depth}
-//                         complexity={style.complexity}
-//                         setToggle={setToggle}
-//                         queries={queries}
-//                         combinedSort={combinedSort}
-//                         // sortByTime={sortByTime}
-//                         // sortByDepth={sortByDepth}
-//                         // sortByComplexity={sortByComplexity}
-//                     />
-//                 </div>
-//             </div>
-//             <div className="chartBox">
-//                 <ChartBox project={setProject} />
-//                 {/* <Chart options={chartData?.options} series={chartData?.series} type="line" /> */}
-//             </div>
-//         </div>
-//     );
-// }
-
-// export default Dashboard;
+    return (
+        <div>
+            <ToolBar
+                projects={allProjects}
+                setSelectedProject={setSelectedProject}
+                projectLoading={loading}
+            />
+            <ProjectView selectedProject={selectedProject} projectLoading={loading} />
+        </div>
+    );
+}
