--- client/components/newDashboard.tsx
@@ -1,53 +1,0 @@
-import React, { useState, useEffect } from 'react';
-// import { gql } from 'apollo-server-express';
-import { useQuery, gql } from '@apollo/client';
-// import { useQuery } from '@apollo/client';
-import ToolBar from './ToolBar';
-import ProjectView from './ProjectView';
-import { useAuth } from '../auth/AuthProvider';
-// import { SelectedProject, Projects } from '../../@types/Interfaces';
-
-const GET_PROJECT_DATA = gql`
-    query getUserData($userId: String!) {
-        user(id: $userId) {
-            projects {
-                name
-                id
-                userID
-                apiKey
-            }
-        }
-    }
-`;
-
-export default function NewDashboard() {
-    /** Bring the user context into this component */
-    const { user } = useAuth();
-
-    /** State requirments for this component */
-    const [selectedProject, setSelectedProject] = useState<Project>();
-    const [allProjects, setAllProjects] = useState<Project[]>();
-
-    /** Send query to get project information for this user */
-    const { data, loading } = useQuery(GET_PROJECT_DATA, {
-        variables: {
-            userId: user!.id,
-        },
-    });
-    useEffect(() => {
-        if (!loading && data) {
-            setAllProjects(data.user.projects);
-        }
-    }, [loading, data]);
-
-    return (
-        <div>
-            <ToolBar
-                projects={allProjects}
-                setSelectedProject={setSelectedProject}
-                projectLoading={loading}
-            />
-            <ProjectView selectedProject={selectedProject} projectLoading={loading} />
-        </div>
-    );
-}
