+++ client/index.tsx
@@ -7,7 +7,7 @@ import { render } from 'react-dom';
 import Signup from './components/Signup';
 import Login from './components/Login';
 // import Dashboard from './components/Dashboard';
-import NewDashboard from './components/NewDashboard';
+import Dashboard from './components/Dashboard';
 import { AuthProvider } from './auth/AuthProvider';
 import Navbar from './components/Navbar';
 import RequireAuth from './components/RequireAuth';
@@ -53,7 +53,7 @@ render(
                             element={
                                 <RequireAuth>
                                     {/**  <Dashboard /> */}
-                                    <NewDashboard />
+                                    <Dashboard />
                                 </RequireAuth>
                             }
                         />
